constrain file ready; bitstream can be created
creation of bitstream works; perhaps the rst_n input of the
topmodul could be removed. For now it is connected to a
low active switch on the AX7015 FPGA board.

now the circuit could be tested on the real hardware (AX7015 board)
together with the Sound modul AN831 which enhance the AX7015 board
i.e. with an 16-bit DAC and line-out output jack. This Sound module
has the chip WM8731 which is powered from the AX7015 board with 3.3V.

Happy testing! Now the Verilog Code of this project should output
the samples of an 1000Hz Sinus Wave in I2S Format to the Sound
module. And the Sound module should create the Sine Wave as analog
signal - to be hearable through the output jack.