FPGA creates a sound through the dac module.
clock_20k is now also in the top module
for easy testing the clock frequenz if it es defined in the constains file
The sound is hearable through the line out of the i2s module. But in my
opinion the sound has a lower frequenz or litte breaks between one wave and
the other perionde of a sine wave. But luckily it makes a digital analog
conversion and it is a sound hearable.