now the I2S DAC module creates right sin signal without cutting the top of the signal.
(the I2C configuation now set a lower speaker loudness, therefore the signal
top is lower, and now not cutted). Also enable the create_clock in the
constrains file. Through this Vivado knows about the created 20 kHz I2C clock,
therefore the 20 kHz clock is synchronized with the 50 MHz clock (syncronized
exactly with the edge of every 2500 cycle of the 50 MHz clock).